pyproject.toml
[tool.poetry]
name = "github-overlord"
version = "0.7.2"
description = "Automatically merge Dependabot PRs in public repositories"
authors = ["Michael Bianco <mike@mikebian.co>"]
readme = "README.md"
keywords = ["github", "dependabot", "automation"]
homepage = "https://github.com/iloveitaly/github-overlord"
include = ["data/*.j2"]

[tool.poetry.dependencies]
python = "^3.11"
pygithub = "^2.4.0"
click = "^8.1.7"
python-decouple-typed = "^3.11.0"
funcy-pipe = ">=0.11.1,<0.15.0"
structlog = ">=24.4,<26.0"
apscheduler = "^3.10.4"
openai = ">=1.52.2,<3.0.0"
pydantic-ai-slim = {extras = ["google"], version = "^0.0.18"}
jinja2 = "^3.1.0"

[tool.poetry.scripts]
github-overlord = "github_overlord:cli"


[tool.isort]
profile = "black"
# helpful when copy/pasting GPT code
float_to_top = true

[tool.poetry.group.dev.dependencies]
isort = ">=5.13.2,<8.0.0"
black = ">=24.10,<26.0"
pyright = "^1.1.386"
pylint = ">=3.3.1,<5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
